backend/requirements.txt
flask>=3.0
flask-session>=0.8
authlib>=1.3
httpx>=0.27
azure-storage-blob>=12.20
python-dotenv>=1.0
pydantic>=2.7
pytest>=8.2
pytest-flask>=1.3
uvicorn[standard]>=0.30
asgiref>=3.8
